backend/pyproject.toml
[project]
name = "adinsights-backend"
version = "0.1.0"
description = "Django backend for multi-tenant advertising insights"
requires-python = ">=3.11"
dependencies = [
  "Django>=5.0.6,<6.0",
  "djangorestframework>=3.15,<4.0",
  "djangorestframework-simplejwt>=5.3,<6.0",
  "django-environ>=0.11,<1.0",
  "psycopg2-binary>=2.9,<3.0",
  "celery>=5.3,<6.0",
  "redis>=5.0,<6.0",
  "cryptography>=46.0.5,<47.0",
  "boto3>=1.34,<2.0",
  "httpx>=0.28,<0.29",
  "croniter>=6.0,<7.0",
  "prometheus-client>=0.20,<0.21",
  "jsonschema>=4.22,<5.0",
  "inflection>=0.5,<0.6",
  "uritemplate>=4.1,<5.0",
  "PyYAML>=6.0,<7.0",
  "Faker>=25.0,<26.0",
  "google-ads>=29.0.0,<30.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2,<9.0",
  "pytest-django>=4.8,<5.0",
  "pytest-cov>=4.1,<5.0",
  "black>=24.4,<25.0",
  "ruff>=0.4,<0.5",
]

[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.build_meta"

[tool.ruff.lint.per-file-ignores]
"backend/config/settings/*.py" = ["F403", "F405"]
